Working using CV2.triangulate
cv2.triangulate was implemented and returns the same values as the two-ray-closest-approach method within 0.05 inches. I needed to fix the R and t values going into the projection matrix calculation because they were not the proper convention. However, cv2.triangulate does not output the reprojection error which is needed to the bundle adjustment, so I do not need this change.